code review: changed 'cpp-c-headers' to old 'c-headers', fixes for Visual Studio compiler 2015 (preview)

+++ src/AnsiRandomNumberGenerator.hpp
@@ -23,7 +23,7 @@
 #ifndef ANSIRANDOMNUMBERGENERATOR_H
 #define ANSIRANDOMNUMBERGENERATOR_H
 
-#include <ctime>
+#include <time.h>
 #include "ARandomNumberGenerator.hpp"
 
 namespace ea
@@ -47,7 +47,7 @@ namespace ea
 
 			AnsiRandomNumberGenerator()
 			{
-				srand(time(nullptr));
+				srand((unsigned int)time(nullptr));
 			}
 
 			~AnsiRandomNumberGenerator() {}
